@@ -1,11 +1,11 @@
 name: flutter_arc_speed_dial
 description: A bunch of FloatingActionButtons to consist speed dial in arc shape
-version: 1.0.3
+version: 2.0.0
 author: May Lau<may.lau.dev@gmail.com>
 homepage: https://github.com/MayLau-CbL/ArcSpeedDial/tree/master/flutter_arc_speed_dial
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
